backend/requirements.txt
# ClipGenius - Backend Dependencies
# 100% FREE - Uses Ollama for AI (local)
fastapi
uvicorn[standard]
python-multipart
yt-dlp
openai-whisper
ffmpeg-python
sqlalchemy
pydantic
python-dotenv
aiosqlite
httpx

# AI Reframe - Face tracking
mediapipe
opencv-python
numpy

# Authentication
python-jose[cryptography]
passlib[bcrypt]
bcrypt

# Payment integrations (optional)
stripe
# mercadopago  # Uncomment for MercadoPago support

# Google Drive integration
google-api-python-client
google-auth-httplib2
google-auth-oauthlib

# Social media integrations (optional)
# tiktok-api
# instagrapi
